Update to v2.61.0 of Zitadel API and a few rebuild script fixes.
Since the last release of this project a lot of new protos have
been added to the zitadel project. The previous rebuild script
used to generate the Elixir SDK in this project relied on the
proto structure being flat and all in one directory.

This change introduces a few changes to `bin/rebuild.sh` that are
needed to compile the protos to Elixir:

- Use a temp directory paired with git sparse checkout to
  download proto files in the download_proto_files function
  (the previous approach would have been more complicated to
  implement recursively)
- Set ERL_AFLAGS environment variable to set standard io_encoding
  to latin1 to fix apparent regression introduced in Erlang 26
  that prevented several of the protos from being compiled
  without the fix

A few changes were also made to make the rebuild script easier
to maintain:

- Move to bash to use pushd/popd
- Separate utility functions into a separate file
- Rename download_files to download_proto_files